add scheduling for eviction check
